Added baseline functionality for user profile
Creates a seperate module for the user_profile, which has a one to one
relationship with the user acount. Uses signals to automatically create
a user profile whenever an account is created, functionally extending
the user account with the additional fields.

used https://www.youtube.com/watch?v=vMVOeBULQZA for reference